requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
autopep8==1.7.1
beautifulsoup4==4.11.1
billiard==4.2.0
celery==5.4.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
distlib==0.3.5
Django==5.0.6
django-allauth==0.54.0
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.8.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.7
MarkupSafe==2.1.1
MouseInfo==0.1.3
msgpack==1.0.4
oauthlib==3.2.2
Pillow==9.5.0
platformdirs==2.5.2
prompt_toolkit==3.0.47
psycopg2==2.9.5
PyAutoGUI==0.9.53
pycodestyle==2.9.1
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytweening==1.0.4
pytz==2023.3.post1
redis==5.0.7
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.3.2.post1
sqlparse==0.5.0
tkPDFViewer==0.1
toml==0.10.2
tomli==2.0.1
twilio==7.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.3
vine==5.1.0
virtualenv==20.16.5
wcwidth==0.2.13
Werkzeug==2.2.2
whitenoise==6.5.0
wikipedia==1.4.0
wrapt==1.14.1
